+++ packages/backend-app/src/private-match-make-polling.ts
@@ -17,7 +17,7 @@ import { WebsocketAPIEvent } from "./lambda/websocket-api-event";
 import { WebsocketAPIResponse } from "./lambda/websocket-api-response";
 import { parsePrivateMatchMakePolling } from "./request/private-match-make-polling";
 import { cloudNotPrivateMatchMake } from "./response/cloud-not-private-match-make";
-import { createBattleStart } from "./response/create-battle-start";
+import { createBattleStart } from "./response/battle-start";
 import { invalidRequestBodyError } from "./response/invalid-request-body-error";
 import { rejectPrivateMatchEntry } from "./response/reject-private-match-entry";
